+++ Cargo.toml
@@ -12,7 +12,6 @@ readme = "README.md"
 [dependencies]
 windows_macros = { path = "crates/macros",  version = "0.3.1" }
 gen = { package = "windows_gen", path = "crates/gen",  version = "0.3.1" }
-winmd = { package = "windows_winmd", path = "crates/winmd",  version = "0.3.1" }
 const-sha1 = "0.2"
 
 [build-dependencies]
