+++ crates/gen/Cargo.toml
@@ -7,8 +7,7 @@ license = "MIT OR Apache-2.0"
 description = "Code generation for the windows crate"
 
 [dependencies]
-windows_gen_macros = { path = "macros",  version = "0.3.1" }
-winmd = { package = "windows_winmd", path = "../winmd",  version = "0.3.1" }
+macros = { package = "windows_gen_macros", path = "macros",  version = "0.3.1" }
 quote = "1.0"
 syn = "1.0"
 proc-macro2 = "1.0"
