+++ crates/gen/src/lib.rs
@@ -18,7 +18,6 @@ mod required_interface;
 mod signature;
 mod r#struct;
 mod to_snake;
-mod traits;
 mod r#type;
 mod type_definition;
 mod type_guid;
@@ -26,6 +25,7 @@ mod type_limits;
 mod type_name;
 mod type_namespaces;
 mod type_tree;
+pub mod winmd;
 
 pub use callback::*;
 pub use class::*;
@@ -48,7 +48,6 @@ pub use r#type::*;
 pub use required_interface::*;
 pub use signature::*;
 pub use to_snake::*;
-pub use traits::*;
 pub use type_definition::*;
 pub use type_guid::*;
 pub use type_limits::*;
