+++ crates/gen/src/winmd/assembly_ref.rs
