+++ crates/gen/src/winmd/attribute.rs
