+++ crates/gen/src/winmd/blob.rs
@@ -1,5 +1,4 @@
-use crate::*;
-
+use super::*;
 use std::convert::TryInto;
 
 pub struct Blob {
