+++ crates/gen/src/winmd/constant.rs
