+++ crates/gen/src/winmd/constant_value.rs
