+++ crates/gen/src/winmd/element_type.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::traits::Decode;
 
 #[derive(Debug)]
 pub enum ElementType {
