+++ crates/gen/src/winmd/flags.rs
