+++ crates/gen/src/winmd/impl_map.rs
