+++ crates/gen/src/winmd/method_def.rs
