+++ crates/gen/src/winmd/module_ref.rs
