+++ crates/gen/src/winmd/nested_class.rs
