+++ crates/gen/src/winmd/type_def.rs
