+++ crates/gen/src/winmd/type_spec.rs
