+++ crates/gen/src/winmd/workspace.rs
