+++ crates/macros/Cargo.toml
@@ -11,7 +11,6 @@ proc-macro = true
 
 [dependencies]
 gen = { package = "windows_gen", path = "../gen", version = "0.3.1" }
-winmd = { package = "windows_winmd", path = "../winmd",  version = "0.3.1" }
 syn = "1.0"
 quote = "1.0"
 squote = "0.1.2"
