--- crates/winmd/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "windows_winmd"
-version = "0.3.1"
-authors = ["Microsoft"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-description = "Winmd parser the windows crate"
-
-[dependencies]
-macros = { package = "windows_winmd_macros", path = "macros",  version = "0.3.1" }
