--- crates/winmd/macros/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "windows_winmd_macros"
-version = "0.3.1"
-authors = ["Microsoft"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-description = "Macro helpers for the windows_winmd crate"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "1.0", features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
